Always reset the dependent select value on change of the parent
This change addresses the necessity to not submit a value for the dependent in case it is inactive/aria-disabled. The same could be achieved by using the HTML attribute `disabled` but we want to keep visible dependents keyboard-focusable with `aria-disabled` so the selected value has to be cleared in JS to avoid false form submit values.