Always reset the dependent select value on change of the parent (#2)
This change addresses the necessity to omit the dependent `<select>`'s
value from form submits when the dependent is inactive.

The same could be achieved by using the HTML attribute `disabled` (no
value submitted), but we want to keep visible dependents
keyboard-focusable[^1], so the selected value has to be cleared in JS to
avoid false form submit values. This may change with #3.

The same could also be achieved by always clearing all options from the
dependent `<select>` on `change` of the parent, but in case the
dependent is always visible, that would necessitate extra logic to
identify and keep the placeholder option (if present). It is easier to
clear the programmatic value without touching the DOM until it needs to
be updated.

[^1]: For accessibility reasons. If the dependent is always visible
(configurable option of the component), it should be perceivable without
vision, i.e. via screenreader output. We're using `aria-disabled` to
communicate the disabled state, which, ironically, can only be
communicated if the element remains focusable.